Don't hardcode path to google/protobuf/descriptor.proto
Despite the ability to specify multiple include directories
in `config.Include.After`, the `/usr/local/bin` path to
`google/protobuf/descriptor.proto` is hardcoded in
`func (*descriptorSet).marshalTo()`.

Fix by introducing a function to search for this file,
and providing the path found to newDesceriptorSet.

This (together with the addition of `/usr/include` path to
`config.Include.After`) makes it possible to use protobuilder
with distro-installed protoc (e.g. with Debian/Ubuntu it is
`apt install libprotobuf-dev protobuf-compiler`).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>